backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
--find-links https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
numpy
opencv-python
Pillow